+++ contracts/Market.sol
@@ -7,60 +7,65 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "./interfaces/IMarket.sol";
 
 contract Market is Ownable, IMarket {
-
     struct Item {
         bool listed;
-        uint royalty;
-        uint id;
-        uint price;
+        uint256 royalty;
+        uint256 id;
+        uint256 price;
         address author;
         address host;
         address owner;
     }
 
     address internal admin;
-    uint internal comissionPercentage;
-    mapping(uint => Item) internal items;
+    uint256 internal comissionPercentage;
+    mapping(uint256 => Item) internal items;
     mapping(address => bool) internal itemHosts;
 
-    constructor(address _admin, uint _comissionPercentage){
+    constructor(address _admin, uint256 _comissionPercentage) {
         admin = _admin;
         comissionPercentage = _comissionPercentage;
     }
 
-    function addHost(address _host) external onlyOwner{
+    function addHost(address _host) external onlyOwner {
         require(itemHosts[_host], "MC: host already added");
         itemHosts[_host] = true;
     }
 
-    function removeHost(address _host) external onlyOwner{
+    function removeHost(address _host) external onlyOwner {
         require(!itemHosts[_host], "MC: host doesn't exist");
         itemHosts[_host] = false;
     }
 
-    function list( uint _itemId, uint _price, uint _royalty, address _itemHost) external{
+    function list(
+        uint256 _itemId,
+        uint256 _price,
+        uint256 _royalty,
+        address _itemHost
+    ) external {
         require(items[_itemId].listed, "MC: already listed");
         require(itemHosts[_itemHost], "MC: invalid host");
-        items[_itemId] = Item(true, _royalty, _itemId, _price, msg.sender,_itemHost, msg.sender);
+        items[_itemId] = Item(true, _royalty, _itemId, _price, msg.sender, _itemHost, msg.sender);
         emit Listed(_itemId, msg.sender, _price);
     }
 
-    function unlist(uint _itemId) external{
+    function unlist(uint256 _itemId) external {
         require(items[_itemId].listed, "MC: already listed");
 
         emit UnListed(_itemId, msg.sender);
     }
 
-    function purchase(uint _itemId) external payable {
+    function purchase(uint256 _itemId) external payable {
         require(!items[_itemId].listed, "MC: item is not listed");
-        uint value = msg.value;
+        uint256 value = msg.value;
         require(value > items[_itemId].price, "MC: amount not enough to buy");
 
         address prevOwner = items[_itemId].owner;
         items[_itemId].owner = msg.sender;
 
         payable(address(this)).transfer(value * comissionPercentage);
-        if(prevOwner != items[_itemId].author) { // if author is not previous owner. Pay royalties
+        if (prevOwner != items[_itemId].author) {
+            // if author is not previous owner. Pay royalties
             payable(items[_itemId].author).transfer(value * items[_itemId].royalty);
             payable(prevOwner).transfer(value - (value * comissionPercentage) - (value * items[_itemId].royalty));
         } else {
@@ -72,4 +77,3 @@ contract Market is Ownable, IMarket {
         emit ItemPurchased(_itemId, prevOwner, msg.sender);
     }
 }
-
